% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COSMIC_v3.0.R
\docType{data}
\name{COSMIC_v3.0}
\alias{COSMIC_v3.0}
\title{Mutational signatures data from COSMIC, Catalogue Of Somatic Mutations In
Cancer (v3.0 - May 2019)}
\format{
A list with one element \code{signature}, with the
same structure as \code{\link{signature}}, except that
sub-element GRCh37 does contain \code{SBS192}.
}
\source{
\url{https://cancer.sanger.ac.uk/signatures/}.
}
\usage{
COSMIC_v3.0
}
\description{
Mutational signatures data from COSMIC, Catalogue Of Somatic Mutations In
Cancer (v3.0 - May 2019)
}
\section{Remark}{

The signatures are all genome signatures.
See \url{https://cancer.sanger.ac.uk/signatures/}.
}

\examples{
sbs96_sig_v3.0 <- COSMIC_v3.0$signature$GRCh37$SBS96

}
\keyword{datasets}
