% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COSMIC_v3.1.R
\docType{data}
\name{COSMIC_v3.1}
\alias{COSMIC_v3.1}
\title{Mutational signatures data from COSMIC, Catalogue Of Somatic Mutations In
Cancer (v3.1 - June 2020)}
\format{
A list with one element \code{signature}, with the
same structure as \code{\link{signature}}.
}
\source{
\url{https://cancer.sanger.ac.uk/signatures/}.
}
\usage{
COSMIC_v3.1
}
\description{
Mutational signatures data from COSMIC, Catalogue Of Somatic Mutations In
Cancer (v3.1 - June 2020)
}
\section{Remark}{

The signatures are all genome signatures.
See \url{https://cancer.sanger.ac.uk/signatures/}.
}

\examples{
sbs96_sig_v3.1 <- COSMIC_v3.1$signature$GRCh37$SBS96

}
\keyword{datasets}
