% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COSMIC_v3.3.R
\docType{data}
\name{COSMIC_v3.3}
\alias{COSMIC_v3.3}
\title{Mutational signatures data from COSMIC, Catalogue Of Somatic Mutations In
Cancer (v3.3 - June 2022)}
\format{
A list with two elements, \code{signature} and \code{etiology}.
\itemize{
\item \code{signature} is a list with the same structure as
\code{\link{signature}}.
\item \code{etiology} is a list with the same structure as
\code{\link{etiology}}.
}
}
\source{
\url{https://cancer.sanger.ac.uk/signatures/}.
}
\usage{
COSMIC_v3.3
}
\description{
Mutational signatures data from COSMIC, Catalogue Of Somatic Mutations In
Cancer (v3.3 - June 2022)
}
\section{Remark}{

The signatures are all genome signatures.
See \url{https://cancer.sanger.ac.uk/signatures/}.
}

\examples{
sbs96_sig_v3.3 <- COSMIC_v3.3$signature$GRCh37$SBS96

}
\keyword{datasets}
