% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COSMIC_v3.0.R
\docType{data}
\name{COSMIC_v3.0}
\alias{COSMIC_v3.0}
\title{Mutational signatures data from COSMIC, the Catalogue Of Somatic Mutations In
Cancer, (v3.0 - May 2019)}
\format{
A list with one element \code{signature}.
\itemize{
\item \code{signature} is a list with the elements:
\itemize{
\item \code{SBS96}: Strand-agnostic single-base substitutions in trinucleotide context.
\item \code{DBS78}: Strand-agnostic doublet-base substitutions.
\item \code{ID}: Strand-agnostic indels.
}
}
}
\source{
\url{https://cancer.sanger.ac.uk/signatures/downloads/}.
}
\usage{
COSMIC_v3.0
}
\description{
Mutational signatures data from COSMIC, the Catalogue Of Somatic Mutations In
Cancer, (v3.0 - May 2019)
}
\section{Remark}{

\code{SBS96} and \code{DBS78} signatures are from Human \strong{GRCh38} reference genome.
\code{SBS192} and \code{ID} signatures are from Human \strong{GRCh37} reference genome.
See \link[ICAMS]{CatalogRowOrder} for the classification of mutation types.
}

\examples{
# As the abundances of the source sequence of the mutations vary between genome
# and exome and between species, users can use ICAMS to do the transformations.
if (!requireNamespace("ICAMS", quietly = TRUE)) {
  install.packages("ICAMS")
}
library(ICAMS)
SBS96_sigs_GRCh38_genome <- COSMIC_v3.0$signature$SBS96

# Transform SBS96 GRCh38 genome signatures to GRCh38 exome signatures
SBS96_sigs_GRCh38_exome <- TransformCatalog(catalog = SBS96_sigs_GRCh38_genome,
                                            target.ref.genome = "GRCh38",
                                            target.region = "exome")

# Transform SBS96 GRCh38 genome signatures to GRCh37 genome signatures
SBS96_sigs_GRCh37_genome <- TransformCatalog(catalog = SBS96_sigs_GRCh38_genome,
                                             target.ref.genome = "GRCh37",
                                             target.region = "genome")

# Transform SBS96 GRCh38 genome signatures to mm10 genome signatures
SBS96_sigs_mm10_genome <- TransformCatalog(catalog = SBS96_sigs_GRCh38_genome,
                                           target.ref.genome = "mm10",
                                           target.region = "genome")
}
\keyword{datasets}
