% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COSMIC_v3.0.R
\docType{data}
\name{COSMIC_v3.0}
\alias{COSMIC_v3.0}
\title{Mutational signatures data from COSMIC, Catalogue Of Somatic Mutations In
Cancer, (v3.0 - May 2019)}
\format{
A list with one element \code{signature}, with the
same structure as \code{\link{signature}}, except that
sub-element GRCh37 does contain \code{SBS192}.
}
\source{
\url{https://cancer.sanger.ac.uk/signatures/}.
}
\usage{
COSMIC_v3.0
}
\description{
Mutational signatures data from COSMIC, Catalogue Of Somatic Mutations In
Cancer, (v3.0 - May 2019)
}
\section{Remark}{

The signatures are all genome signatures. See examples below for transforming
genome signatures to exome signatures. \cr See \link[ICAMS]{CatalogRowOrder}
in package ICAMS for the classification of mutation types.
}

\examples{
\dontrun{
# As the abundances of the source sequence of the mutations vary between genome
# and exome, users can use package ICAMS to do the transformations.
if (!requireNamespace("ICAMS", quietly = TRUE)) {s
  install.packages("ICAMS")
}
library(ICAMS)
SBS96_sig_GRCh37_genome <- COSMIC_v3.0$signature$GRCh37$SBS96

# Transform SBS96 GRCh37 genome signatures to GRCh37 exome signatures
SBS96_sig_GRCh37_exome <- TransformCatalog(
  catalog = SBS96_sig_GRCh37_genome,
  target.ref.genome = "GRCh37",
  target.region = "exome"
)
}
}
\keyword{datasets}
