% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COSMIC_v3.1.R
\docType{data}
\name{COSMIC_v3.1}
\alias{COSMIC_v3.1}
\title{Mutational signatures data from COSMIC, the Catalogue Of Somatic Mutations In
Cancer, (v3.1 - June 2020)}
\format{
A list with one element \code{signature}.
\itemize{
\item \code{signature} is a list with one element:
\itemize{
\item \code{genome}: A list with the elements:
\itemize{
\item \code{SBS96}: Strand-agnostic single-base substitutions in trinucleotide context.
\item \code{SBS192}: Transcriptionally stranded single-base substitutions in trinucleotide context.
\item \code{DBS78}: Strand-agnostic doublet-base substitutions.
\item \code{ID}: Strand-agnostic indels.
}
}
}
}
\source{
Files downloaded from
\url{https://cancer.sanger.ac.uk/signatures/downloads/}, 2021 Oct and
saved in \code{data-raw/COSMIC_v3.1/data/}. \cr Populated by
\code{data-raw/COSMIC_v3.1/code/generate_COSMIC_v3.1_genome_sigs.R}.
}
\usage{
COSMIC_v3.1
}
\description{
Mutational signatures data from COSMIC, the Catalogue Of Somatic Mutations In
Cancer, (v3.1 - June 2020)
}
\section{Remark}{

The signatures are all from Human GRCh37 reference genome.
}

\examples{
COSMIC_v3.1_SBS96.sigs <- COSMIC_v3.1$signature$genome$SBS96
}
\keyword{datasets}
