% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COSMIC_v3.5.R
\docType{data}
\name{COSMIC_v3.5}
\alias{COSMIC_v3.5}
\title{Mutational signatures data from COSMIC, Catalogue Of Somatic Mutations In
Cancer (v3.5 - November 2025)}
\format{
A list with two elements, \code{signature} and \code{etiology}.
\itemize{
\item \code{signature} is a list with the same structure as
\code{\link{signature}}.
\item \code{etiology} is a list with the same structure as
\code{\link{etiology}}. The etiology information is deprecated
and is no longer being updated. It is not versioned at the COSMIC
website and may not reflect signatures added in recent COSMIC
releases. For authoritative, up-to-date etiologies consult
\url{https://cancer.sanger.ac.uk/signatures/} directly.
}
}
\source{
\url{https://cancer.sanger.ac.uk/signatures/}.
}
\usage{
COSMIC_v3.5
}
\description{
Mutational signatures data from COSMIC, Catalogue Of Somatic Mutations In
Cancer (v3.5 - November 2025)
}
\section{Remark}{

The signatures are all genome signatures.
See \url{https://cancer.sanger.ac.uk/signatures/}.
}

\examples{
sbs96_sig_v3.5 <- COSMIC_v3.5$signature$GRCh37$SBS96

}
\keyword{datasets}
