% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosmicsig.R
\docType{package}
\name{cosmicsig}
\alias{cosmicsig}
\title{cosmicsig: A package of mutational signatures data from COSMIC (the Catalogue
Of Somatic Mutations In Cancer) website
\url{https://cancer.sanger.ac.uk/signatures/}}
\source{
COSMIC mutational signatures data were downloaded from
\url{https://cancer.sanger.ac.uk/signatures/downloads/}.
}
\description{
This is a data package with 2 main package variables:
\code{\link{signature}} and \code{\link{etiology}}.
}
\details{
There are also two functions for handling COSMIC signatures:
\itemize{
\item \code{\link{get_etiology}}
\item \code{\link{SBS96_ID_to_SBS192_ID}}
}
}
