% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etiology.R
\docType{data}
\name{etiology}
\alias{etiology}
\title{List of mutational signatures' proposed etiology summarized from COSMIC,
Catalogue Of Somatic Mutations In Cancer (v3.5 - November 2025)}
\format{
A list with the elements:
\itemize{
\item SBS96
\item DBS78
\item ID
}

Each element is a single-column matrix with rownames being the signature
IDs and values being a short string describing the proposed
etiology.

In general use \code{\link{get_etiology}}, which handles new
signatures do not have an element in \code{etiology}.
}
\source{
\url{https://cancer.sanger.ac.uk/signatures/}.
}
\usage{
etiology
}
\description{
List of mutational signatures' proposed etiology summarized from COSMIC,
Catalogue Of Somatic Mutations In Cancer (v3.5 - November 2025)
}
\examples{
sbs96_etiology <- etiology$SBS96
}
\keyword{datasets}
