% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_etiology.R
\name{get_etiology}
\alias{get_etiology}
\title{Get the proposed etiology of COSMIC signature}
\usage{
get_etiology(mutation_type, sig_id, cosmic_version = "3.2")
}
\arguments{
\item{mutation_type}{Character string, one of "SBS96", "SBS192", "DBS78",
"ID".}

\item{sig_id}{Character vector with signature ids, e.g. \code{c("SBS3",
"SBS5")}.}

\item{cosmic_version}{COSMIC signature version.}
}
\value{
A character vector of the same length as \code{sig_id},
each element of which is the etiology of the
corresponding signature, if known, or else the empty string.
}
\description{
Get the proposed etiology of COSMIC signature
}
\examples{
get_etiology(mutation_type = "ID", sig_id = c("ID1", "foo", "ID3"))

}
