% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rare_signatures.R
\name{rare_signatures}
\alias{rare_signatures}
\title{Return a character vector of the names of rare SBS96 signatures}
\usage{
rare_signatures()
}
\value{
A character vector of the names of rare SBS96 signatures.
}
\description{
Return a character vector of the names of rare SBS96 signatures
}
\examples{
rare_sigs <- rare_signatures()
}
