% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\docType{data}
\name{signature}
\alias{signature}
\title{Mutational signatures data from COSMIC, the Catalogue Of Somatic Mutations In
Cancer (v3.2 - March 2021)}
\format{
A list with the following elements:
*\code{SBS96}: Strand-agnostic single-base substitutions in trinucleotide context.
*\code{SBS192}: Transcriptionally stranded single-base substitutions in trinucleotide context.
*\code{DBS78}: Strand-agnostic doublet-base substitutions.
*\code{ID}: Strand-agnostic indels.
}
\source{
\url{https://cancer.sanger.ac.uk/signatures/downloads/}.
}
\usage{
signature
}
\description{
Mutational signatures data from COSMIC, the Catalogue Of Somatic Mutations In
Cancer (v3.2 - March 2021)
}
\section{Remark}{

The signatures are all from Human GRCh37 reference genome.
See \link[ICAMS]{CatalogRowOrder} for the classification of mutation types.
}

\section{Note}{

SBS10c, SBS10d, SBS91, SBS92, SBS93, SBS94 (total 6) new SBS signatures were
added in COSMIC v3.2. See the news from COSMIC release for more details
\url{https://cosmic-blog.sanger.ac.uk/cosmic-mutational-signatures-release-v3-2/}
}

\examples{
SBS96_sigs <- signature$genome$SBS96
}
\keyword{datasets}
