% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\docType{data}
\name{signature}
\alias{signature}
\title{Mutational signatures data from COSMIC, the Catalogue Of Somatic Mutations In
Cancer (v3.2 - March 2021)}
\format{
A list with a single element, \code{genome}, which is a
list containing:
\describe{
\item{SBS96}{Strand-agnostic single-base substitutions in trinucleotide context.}
\item{SBS192}{Transcriptionally stranded single-base substitutions in trinucleotide context.}
\item{DBS78}{Strand-agnostic doublet-base substitutions.}
\item{ID}{Strand-agnostic indels.}
}
}
\source{
Files downloaded from
\url{https://cancer.sanger.ac.uk/signatures/downloads/}, 2021 Sep and
saved in \code{data-raw/COSMIC_v3.2/data/}. \cr Populated by
\code{data-raw/COSMIC_v3.2/code/generate_COSMIC_v3.2_genome_sigs.R}.
}
\usage{
signature
}
\description{
Mutational signatures data from COSMIC, the Catalogue Of Somatic Mutations In
Cancer (v3.2 - March 2021)
}
\section{Remark}{

The signatures are all from Human GRCh37 reference genome.
See \link[ICAMS]{CatalogRowOrder} for the classification of mutation types.
}

\section{Note}{

SBS10c, SBS10d, SBS91, SBS92, SBS93, SBS94 (total 6) new SBS signatures were
added in COSMIC v3.2. See the news from COSMIC release for more details
\url{https://cosmic-blog.sanger.ac.uk/cosmic-mutational-signatures-release-v3-2/}
}

\examples{
SBS96_sigs <- signature$genome$SBS96
}
\keyword{datasets}
