% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\docType{data}
\name{signature}
\alias{signature}
\title{Mutational signatures data from COSMIC, the Catalogue Of Somatic Mutations In
Cancer (v3.2 - March 2021)}
\format{
A list with the following elements:
\itemize{
\item \code{GRCh37}: Homo sapiens (human) genome assembly GRCh37.
\itemize{
\item \code{SBS96}: Strand-agnostic single-base substitutions in trinucleotide context.
\item \code{SBS192}: Transcriptionally stranded single-base substitutions in trinucleotide context.
\item \code{DBS78}: Strand-agnostic doublet-base substitutions.
\item \code{ID}: Strand-agnostic indels.
}
\item \code{GRCh38}: Homo sapiens (human) genome assembly GRCh38.
\itemize{
\item \code{SBS96}: Strand-agnostic single-base substitutions in trinucleotide context.
\item \code{DBS78}: Strand-agnostic doublet-base substitutions.
}
\item \code{mm9}: Mus musculus (house mouse) genome assembly mm9.
\itemize{
\item \code{SBS96}: Strand-agnostic single-base substitutions in trinucleotide context.
\item \code{DBS78}: Strand-agnostic doublet-base substitutions.
}
\item \code{mm10}: Mus musculus (house mouse) genome assembly mm10.
\itemize{
\item \code{SBS96}: Strand-agnostic single-base substitutions in trinucleotide context.
\item \code{DBS78}: Strand-agnostic doublet-base substitutions.
}
\item \code{rn6}: Rattus norvegicus (Norway rat) genome assembly rn6.
\itemize{
\item \code{SBS96}: Strand-agnostic single-base substitutions in trinucleotide context.
\item \code{DBS78}: Strand-agnostic doublet-base substitutions.
}
}
}
\source{
\url{https://cancer.sanger.ac.uk/signatures/}.
}
\usage{
signature
}
\description{
Mutational signatures data from COSMIC, the Catalogue Of Somatic Mutations In
Cancer (v3.2 - March 2021)
}
\section{Remark}{

See examples below for transforming genome signatures to exome signatures. \cr
See \link[ICAMS]{CatalogRowOrder} in package ICAMS for the classification of mutation types.
}

\section{Note}{

SBS10c, SBS10d, SBS91, SBS92, SBS93, SBS94 (total 6) new SBS signatures were
added in COSMIC v3.2. See the news from COSMIC release for more details
\url{https://cosmic-blog.sanger.ac.uk/cosmic-mutational-signatures-release-v3-2/}
}

\examples{
\dontrun{
# As the abundances of the source sequence of the mutations vary between genome
# and exome, users can use package ICAMS to do the transformations.
if (!requireNamespace("ICAMS", quietly = TRUE)) {
  install.packages("ICAMS")
}
library(ICAMS)
SBS96_sig_GRCh37_genome <- signature$GRCh37$SBS96

# Transform SBS96 GRCh37 genome signatures to GRCh37 exome signatures
SBS96_sig_GRCh37_exome <- TransformCatalog(catalog = SBS96_sig_GRCh37_genome,
                                           target.ref.genome = "GRCh37",
                                           target.region = "exome")
}
}
\keyword{datasets}
